moves ggml-vocab.bin into test folder where it's used.
It appears this file is only used during tests as of now.
Removing it from the model folder makes it more flexible
for how users are loading their model data into the project
(e.g. are they using a docker bind-mounts, are they using
symlinks, are they DLing models directly into this folder?)

By moving this, the instructions for getting started can be
safely simplified to:

$  rm models/.gitkeep
$  rm -r models
$  ln -s /mnt/c/ai/models/LLaMA $(pwd)/models

I think it's a good idea because the model files are quite large, and
be useful across multiple projects so symlinks shine in this use case
without creating too much confusion for the onboardee..