documentation: reflow the readme to make following the setup more clear
This minor (though time consuming) change,

1) Moves the models/gmml-vocab.bin file into the test folder.

2) changes the order in which information is presented to the user

3) recommends using symlinks to link training data into the right place in the repo

4) adds some clarification around the importance of the model weights

1 is handy because it enables 'automation' towards 3, e.g. the
command rm -r models/ can safely be symlinked into the models folder
and the commands to do so are clearly listed and described in the README.md

2 is ultimately the only important aspect of this change.  The readme
currently must be read in full by the user, cached, and then returned to
in order to follow along with all the steps in the documentation.

3 is (I think) handy because these files are pretty huge and not exclusive
to this repo.  Symlinks shine in this as that many symlinks can be created
across multiple projects and all point to the same source location.
If researchers were copying/ pasting these to each project, it would get
out of hand fast I think.

4 seems valuable, the AI world looks really opaque to people just getting started.
I did my best to be accurate with my statements in the hops that it makes
it more possible for humans to become more aware of this technology and
what's happening to the internet and the world.